refactor(carousel): drop cjs; convert to esm only
BREAKING CHANGE: Remove file path imports (eg, 'packagename/react')
BREAKING CHANGE: Remove cjs build
