+++ packages/carousel/package.json
@@ -4,15 +4,14 @@
   "description": "Carousel UI Component for the Pluralsight Design System",
   "license": "Apache-2.0",
   "repository": "pluralsight/design-system",
-  "main": "dist/cjs/index.js",
+  "main": "dist/esm/index.js",
   "module": "dist/esm/index.js",
   "publishConfig": {
     "access": "public"
   },
   "scripts": {
-    "build": "run-s build:cjs build:esm build:css",
-    "build:cjs": "tsc --project tsconfig.build.json --module commonjs --target es5 --outDir dist/cjs",
-    "build:css": "build-css --useGlamor -i dist/cjs/css/index.js",
+    "build": "run-s  build:esm build:css",
+    "build:css": "build-css --useGlamor -i dist/esm/css/index.js",
     "build:esm": "tsc --project tsconfig.build.json --module es2015 --target es5 --outDir dist/esm",
     "build:watch": "yarn build:esm --watch",
     "storybook": "start-storybook -p 6006",
@@ -20,7 +19,9 @@
     "test:watch": "yarn test --watchAll"
   },
   "gitHead": "7418883d96a76c59a99f86c2b16516ec752bf913",
+  "sideEffects": false,
   "style": "dist/index.css",
+  "type": "module",
   "types": "dist/esm/index.d.ts",
   "dependencies": {
     "@pluralsight/ps-design-system-core": "^6.9.3",
