+++ packages/carousel/src/js/__specs__/utils.spec.ts
@@ -1,4 +1,4 @@
-import * as utils from '../utils'
+import * as utils from '../utils.js'
 
 describe('utils', () => {
   describe('#chunk', () => {
