+++ packages/carousel/src/react/use-unique-id.ts
@@ -1,6 +1,6 @@
 import { useLayoutEffect, useState } from 'react'
 
-import { uniqueId } from '../js/utils'
+import { uniqueId } from '../js/utils.js'
 
 export default function useUniqueId(
   prefix = '',
